requirements.txt
xlwings==0.30.13
pillow==10.2.0
python-pptx==0.6.23
comtypes==1.2.1